backend/requirements.txt
fastapi
uvicorn
requests
pydantic
google-generativeai
python-dotenv
groq
# oumi (Commented out due to build error on Windows)